requirements.txt
Telethon==1.42.0
cryptg==0.5.2
python-dotenv==1.2.2
pytest==8.3.5
